style: fix meet invite page
