+++ src/app/(main)/meet/[meetId]/page.tsx
@@ -7,10 +7,10 @@ import {
   getMyParticipatingMeets,
 } from '@/controllers/meet';
 import { redirect } from 'next/navigation';
-import EditMeetButton from '../EditMeetButton';
-import ParticipantsButton from '../ParticipantsButton';
-import MeetDescription from '../MeetDescription';
-import ShareLinkButton from '../ShareLinkButton';
+import EditMeetButton from '../_component/EditMeetButton';
+import ParticipantsButton from '../_component/ParticipantsButton';
+import MeetDescription from '../_component/MeetDescription';
+import ShareLinkButton from '../_component/ShareLinkButton';
 import { getCurrentUser } from '@/lib/authentication';
 import TimeTable from '@/types/TimeTable';
