+++ src/app/(main)/meet/[meetId]/participate/page.tsx
@@ -1,6 +1,8 @@
 import { getMeet } from '@/controllers/meet';
-import MeetDescription from '../../MeetDescription';
-import ParticipateForm from './ParticipateForm';
+import MeetDescription from '../../_component/MeetDescription';
+import ParticipateForm from './_component/ParticipateForm';
+import { InfoCardWrapper } from '@/components/login/info-card';
+
 export default async function ParticipateMeetPage({
   params,
 }: {
@@ -8,13 +10,14 @@ export default async function ParticipateMeetPage({
 }) {
   const meet = await getMeet(params.meetId);
   return (
-    <>
+    <div className="flex flex-col items-center gap-6">
       <p className="text-xl mt-3">{meet.name}</p>
-      <MeetDescription description={meet.description} />
+      <InfoCardWrapper source="https://raw.githubusercontent.com/Tarikul-Islam-Anik/Animated-Fluent-Emojis/master/Emojis/Objects/Calendar.png" />
+      <p>{meet.description}</p>
       <ParticipateForm
         meetId={params.meetId}
         isProtected={meet.password ? true : false}
       />
-    </>
+    </div>
   );
 }
