+++ src/app/(main)/mypage/edit/page.tsx
@@ -1,5 +1,5 @@
 import { getCurrentUser } from '@/lib/authentication';
-import EditProfileForm from './EditProfileForm';
+import EditProfileForm from './_component/EditProfileForm';
 
 export default async function EditProfilePage() {
   const currentUser = await getCurrentUser();
